Platform/ARM/Morello: Add FVP GOP config and GOP DDR carve-out
The platform GOP frame-buffer must be within the lower 40-bits of the
memory space.
To guarantee this and make sure that frame-buffer memory is always
available we carve out a reservation at the end of the 32-bit DDR
mapping.

The size of this carve-out is controlled by the PCD
gArmMorelloTokenSpaceGuid.PcdPlatformGopBufferSize and a size of 0 means
to disable memory reservation for the platform frame-buffer.

Signed-off-by: Patrik Berglund <patrik.berglund@arm.com>
Change-Id: I85e9f00162408e82fcd518aa2c3aa5ac3ff9a76e