Platform/ARM/Morello: Fix SMBIOS Type 3 (System Enclosure) length
The SMBIOS header for this table was initialised with the wrong
table's size, which was 3 bytes greater than the correct value.
As a result,  the first three characters from the strings, the
"ARM" in "ARM LTD" for the chassis manufacturer, are erroneously
attributed to the table and so parsers interpret the manufacturer
as " LTD".

Signed-off-by: Jessica Clarke <jessica.clarke@cl.cam.ac.uk>
Change-Id: I6acad50b78f9241f1b331aa29e9a1adeff0fac78