Platform/ARM/Morello: Add UIDs to CoreSight devices
In ACPI, multiple instances of an IP block in the same namespace may be
differentiated by the Unique Identifier (_UID) object. This must be
unique across all devices with a common _HID or _CID. _UID objects are
added for all CoreSight devices with multiple instances in Morello, so
that each device can be uniquely identified.

Signed-off-by: Werner Lewis <werner.lewis@arm.com>
Change-Id: I625ff1c623c103a9b34bf1cf4dbb2aa7433c30b7